Keep track of installed fields in Schema. Add install method to fields, move logic from Schema to fields.
